Update whitelist.txt
Removed sites that are no longer existing.
Easybytez and Tusfiles closed down a long time ago. Now the tusfiles domain redirects people to send.now, which, as far as I can see, is a new domain for send.cm, which then became send.com.nl and is now send.now.